♻️ 코어 데이터 이름 변경 및 SaveCoreData 기능 수정
