+++ Alarmi/Alarmi/Sources/CoreData/SaveCoreData.swift
@@ -1,8 +1,8 @@
 //
-//  SaveCoreData.swift
+//  UpdateCoreData.swift
 //  Alarmi
 //
-//  Created by 임 용관 on 2022/07/29.
+//  Created by 임 용관 on 2022/07/30.
 //  Copyright © 2022 MoTe. All rights reserved.
 //
 
@@ -10,33 +10,99 @@ import Foundation
 import CoreData
 import UIKit
 
-func saveCoreData(start: Date, end: Date) -> Bool {
-    
-    // App Delegate 호출
+func saveCoreData(day: Date, keyName: String) -> Bool {
     guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { return false }
-    
-    // App Delegate 내부에 있는 viewContext 호출
-    let context: NSManagedObjectContext = appDelegate.persistentContainer.viewContext
-
-    // managedContext 내부에 있는 entity 호출
-    let entity = NSEntityDescription.entity(forEntityName: "CoreDataDate", in: context)!
-    
-    // entity 객체 생성
-    let object = NSManagedObject(entity: entity, insertInto: context)
-    
-    // 값 설정
-    object.setValue(start, forKey: "start")
-    object.setValue(end, forKey: "end")
-    object.setValue(UUID(), forKey: "id")
+    let managedContext = appDelegate.persistentContainer.viewContext
+    let fetchRequest = NSFetchRequest<NSFetchRequestResult>.init(entityName: "CoreDataDate")
     
     do {
-        // managedContext 내부의 변경사항 저장
-        print("save")
-        try context.save()
+        let result = try managedContext.fetch(fetchRequest)
+        let object = result.last as! NSManagedObject
+        
+        object.setValue(day, forKey: keyName)
+
+        try managedContext.save()
+        
         return true
     } catch let error as NSError {
-        // 에러 발생시
-        print("Could not save. \(error), \(error.userInfo)")
+        print("Could not update. \(error), \(error.userInfo)")
+        return false
+    }
+}
+
+func saveCoreData(period: Int16, keyName: String) -> Bool {
+    guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { return false }
+    let managedContext = appDelegate.persistentContainer.viewContext
+    let fetchRequest = NSFetchRequest<NSFetchRequestResult>.init(entityName: "CoreDataDate")
+    
+    do {
+        let result = try managedContext.fetch(fetchRequest)
+        let object = result.last as! NSManagedObject
+        
+        object.setValue(period, forKey: keyName)
+        
+        try managedContext.save()
+        return true
+    } catch let error as NSError {
+        print("Could not update. \(error), \(error.userInfo)")
+        return false
+    }
+}
+
+func saveCoreData(isCall: Bool, keyName: String) -> Bool {
+    guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { return false }
+    let managedContext = appDelegate.persistentContainer.viewContext
+    let fetchRequest = NSFetchRequest<NSFetchRequestResult>.init(entityName: "CoreDataDate")
+    
+    do {
+        let result = try managedContext.fetch(fetchRequest)
+        let object = result.last as! NSManagedObject
+        
+        object.setValue(isCall, forKey: keyName)
+        
+        try managedContext.save()
+        return true
+    } catch let error as NSError {
+        print("Could not update. \(error), \(error.userInfo)")
+        return false
+    }
+}
+
+func updateToday() -> Bool {
+    // Update 오늘 되도록
+    // 중복 배열 X
+    let day = Date()
+    let keyName = "moteDate"
+    
+    guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { return false }
+    let managedContext = appDelegate.persistentContainer.viewContext
+    let fetchRequest = NSFetchRequest<NSFetchRequestResult>.init(entityName: "CoreDataDate")
+    
+    do {
+        let result = try managedContext.fetch(fetchRequest)
+        guard let object = result.last as? CoreDataDate else { return false }
+
+        if object.moteDate == nil {
+            object.setValue([day], forKey: keyName)
+        } else {
+            guard var curDate = object.moteDate else { return false }
+            
+            guard let lastDay = curDate.last else {return false}
+            let lastDayComponent = Calendar.current.dateComponents([.year, .month, .day], from: lastDay)
+            let todayComponent = Calendar.current.dateComponents([.year, .month, .day], from: Date())
+            if lastDayComponent == todayComponent {
+                print("Fail!")
+                return false
+            } else {
+                print("Success!")
+                curDate.append(day)
+                object.setValue(curDate, forKey: keyName)
+            }
+        }
+        try managedContext.save()
+        return true
+    } catch let error as NSError {
+        print("Could not update. \(error), \(error.userInfo)")
         return false
     }
 }
